Change from int64 to int on armorXCharacterData & backgroundXSkill

+++ internal/armorXCharacterData/interface.go
@@ -4,20 +4,20 @@ import "github.com/proyecto-dnd/backend/internal/domain"
 
 type RepositoryArmorXCharacterData interface {
 	CreateArmorXCharacterData(data domain.ArmorXCharacterData) (domain.ArmorXCharacterData, error)
-	DeleteArmorXCharacterData(id int64) error
-	DeleteByCharacterDataIdArmor(id int64) error
+	DeleteArmorXCharacterData(id int) error
+	DeleteByCharacterDataIdArmor(id int) error
 	GetAllArmorXCharacterData() ([]domain.ArmorXCharacterData, error)
-	GetByIdArmorXCharacterData(id int64) (domain.ArmorXCharacterData, error)
-	GetByCharacterDataIdArmor(id int64) ([]domain.ArmorXCharacterData, error)
+	GetByIdArmorXCharacterData(id int) (domain.ArmorXCharacterData, error)
+	GetByCharacterDataIdArmor(id int) ([]domain.ArmorXCharacterData, error)
 	UpdateArmorXCharacterData(data domain.ArmorXCharacterData) (domain.ArmorXCharacterData, error)
 }
 
 type ServiceArmorXCharacterData interface {
 	CreateArmorXCharacterData(data domain.ArmorXCharacterData) (domain.ArmorXCharacterData, error)
-	DeleteArmorXCharacterData(id int64) error
-	DeleteByCharacterDataIdArmor(id int64) error
+	DeleteArmorXCharacterData(id int) error
+	DeleteByCharacterDataIdArmor(id int) error
 	GetAllArmorXCharacterData() ([]domain.ArmorXCharacterData, error)
-	GetByIdArmorXCharacterData(id int64) (domain.ArmorXCharacterData, error)
-	GetByCharacterDataIdArmor(id int64) ([]domain.ArmorXCharacterData, error)
+	GetByIdArmorXCharacterData(id int) (domain.ArmorXCharacterData, error)
+	GetByCharacterDataIdArmor(id int) ([]domain.ArmorXCharacterData, error)
 	UpdateArmorXCharacterData(data domain.ArmorXCharacterData) (domain.ArmorXCharacterData, error)
 }
