+++ internal/domain/backgroundXSkills.go
@@ -1,7 +1,7 @@
 package domain
 
 type BackgroundXSkills struct {
-	BackgroundXSkills_Id int64 `json:"backgroundxskillsid"`
-	Background_Id        int64 `json:"backgroundid"`
-	Skill_Id             int64 `json:"skillid"`
+	BackgroundXSkills_Id int `json:"backgroundxskillsid"`
+	Background_Id        int `json:"backgroundid"`
+	Skill_Id             int `json:"skillid"`
 }
